Simulation writes are delayed iff triggered from an edge event
If a thread writes to a signal, it is delayed if and only if triggered from an
edge event. Before this change it depended on whether the written signal is a
Reg or Wire. While this matches RTL code, it leads to very confusing semantics
when explicitly specifying events as required for simulation threads.